spravene  metody pre medziscreen po prihlasovani

+++ src/main/java/com/example/AutoskolaDemoWithSecurity/services/DrivingSchoolService.java
@@ -5,12 +5,14 @@ import com.example.AutoskolaDemoWithSecurity.models.databaseModels.DrivingSchool
 import com.example.AutoskolaDemoWithSecurity.models.databaseModels.Relationship;
 import com.example.AutoskolaDemoWithSecurity.models.databaseModels.User;
 import com.example.AutoskolaDemoWithSecurity.models.transferModels.DrivingSchoolDTO;
-import com.example.AutoskolaDemoWithSecurity.models.transferModels.UserConfirmation;
+import com.example.AutoskolaDemoWithSecurity.models.transferModels.UserRelationInfo;
+import com.example.AutoskolaDemoWithSecurity.repositories.CompletedRideRepository;
 import com.example.AutoskolaDemoWithSecurity.repositories.ConfirmationRepository;
 import com.example.AutoskolaDemoWithSecurity.repositories.DrivingSchoolRepository;
 import com.example.AutoskolaDemoWithSecurity.repositories.RelationshipRepository;
 import com.example.AutoskolaDemoWithSecurity.repositories.UserRepository;
 import java.nio.file.AccessDeniedException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.persistence.EntityNotFoundException;
@@ -18,7 +20,6 @@ import javax.persistence.PersistenceException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -45,25 +46,24 @@ public class DrivingSchoolService {
     
     
     public ResponseEntity createDrivingSchool(DrivingSchoolDTO drivingSchoolDTO) {
-            User owner = (userRepository.findByEmail(SecurityContextHolder
-                        .getContext().getAuthentication().getName()).get());
-        System.out.println("owner's email: "+owner.getEmail());
-        
-        DrivingSchool drivingSchool = new DrivingSchool(drivingSchoolDTO);
-        drivingSchool.setOwner(owner);
+        User owner = userRepository.findByEmail(drivingSchoolDTO.getOwnerEmail())
+                .orElseThrow(() -> new EntityNotFoundException("No user with email: "+drivingSchoolDTO.getEmail()));
+        if(owner.getRoles().contains("OWNER")) {
+            DrivingSchool drivingSchool = new DrivingSchool(drivingSchoolDTO);
+            drivingSchool.setOwner(owner);
 
             DrivingSchool school = schoolRepository.save(drivingSchool);
             Relationship r = relationshipRepository.save(new Relationship(school, owner, true));
-
-        if (school == null || r == null) {
-            System.out.println("1 of the values is null, school: "+school+", relationship: "+r);
-            schoolRepository.delete(school);
-            relationshipRepository.delete(r);
-        throw new PersistenceException("Something went wrong, School " + drivingSchool.getName() + " could not be saved");
+            if (school == null || r == null) {
+                schoolRepository.delete(school);
+                relationshipRepository.delete(r);
+                throw new PersistenceException("Something went wrong, School " + drivingSchool.getName() + " could not be saved");
+            }
+            return new ResponseEntity("School succesfully created, wait for", HttpStatus.OK);
         }
-        return new ResponseEntity("School succesfully created", HttpStatus.OK);
-        
+        return new ResponseEntity("This user has no OWNER role", HttpStatus.BAD_REQUEST);
     }
+    
     
     public ResponseEntity getSchoolInfo(int id) throws AccessDeniedException {
         
@@ -76,31 +76,78 @@ public class DrivingSchoolService {
         }
     }
     
+    
     public List<DrivingSchoolDTO> getDrivingSchools() {
         List<DrivingSchool> list = schoolRepository.findAll();
         return list.stream().map(DrivingSchoolDTO::new).collect(Collectors.toList());
     }
     
-    public List<UserConfirmation> viewRequests(int relationID){
+    
+    public List<UserRelationInfo> viewRequests(int relationID){
         DrivingSchool school = relationshipRepository.findById(relationID).get().getDrivingSchool();
         return confirmationRepository.findAllByDrivingSchool(school)
                 .stream()
-                .map(UserConfirmation::new)
+                .map(UserRelationInfo::new)
                 .collect(Collectors.toList());
     }
     
     //ak true - najde relationship - aktivuje ho - vymaze verifikaciu
     //ak false - vymaze verifikaciu, relationship neha, kvoli userovi
-    public String confirmUser(int userRelationID, boolean confirm) {
-        if(confirm) {
-            try {
-                relationshipService.activateRelationship(relationshipRepository.findById(userRelationID).get());
-            } catch(Exception e) {
-                throw new EntityNotFoundException("Wrong relationship ID");
+    public ResponseEntity confirmUser(int userRelationID, boolean confirm) {
+        Relationship relation = relationshipRepository.findById(userRelationID).get();
+        if(relation.getDrivingSchool().getOwner().getEmail()
+                .equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+            
+            if(confirm) {
+                try {
+                    relationshipService.activateRelationship(relation, true);
+                } catch(Exception e) {
+                    throw new EntityNotFoundException("Wrong relationship ID");
+                }
+            } else {
+                relation.setStatus("rejected");
+                relationshipRepository.save(relation);
             }
+            confirmationRepository.deleteByRelation(userRelationID);
+            return new ResponseEntity(confirm ? "User confirmed" : "User rejected", HttpStatus.OK);
         }
-        confirmationRepository.deleteByRelation(userRelationID);
-        return confirm ? "User confirmed" : "User rejected";
+        return new ResponseEntity("ID does not belong to your school", HttpStatus.BAD_REQUEST);
+    }
+    
+    
+    public ResponseEntity completeStudent(int studentRelationID, boolean completed) {
+        Relationship relation = relationshipRepository.findById(studentRelationID).orElseThrow(
+                    () -> new EntityNotFoundException("This relationID does not exist"));
+        System.out.println("com.example.AutoskolaDemoWithSecurity.services.DrivingSchoolService.completeStudent()");
+         if(relation.getDrivingSchool().getOwner().getEmail()
+                .equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+             System.out.println("com.example.AutoskolaDemoWithSecurity.services.DrivingSchoolService.completeStudent()");
+             if(completed) {
+                 System.out.println("com.example.AutoskolaDemoWithSecurity.services.Drivin          TRUE    ");
+                 if(relation.getStatus().equalsIgnoreCase("waiting for exam")) {
+                     relationshipService.changeRelationshipStatus(relation, "completed");
+                 } else {
+                     return new ResponseEntity("This user did not completed rides yet", HttpStatus.BAD_REQUEST);
+                 }
+             } else {
+                 System.out.println("com.example.AutoskolaDemoWithSecurity.services.            FALSE");
+                 relationshipService.changeRelationshipStatus(relation, "not completed");
+             }
+             
+         } else {
+             return new ResponseEntity("ID does not belong to your school", HttpStatus.BAD_REQUEST);
+         }
+        return new ResponseEntity("Student status set as: "+(completed ? "completed" : "not completed"), HttpStatus.OK);
+    }
+    
+    public List<UserRelationInfo> getCompletedUsers(int relationID) {
+        List<UserRelationInfo> list = new ArrayList<UserRelationInfo>();
+        list.addAll(relationshipRepository.findAllByDrivingSchoolAndStatus(
+                relationshipRepository.findById(relationID).get().getDrivingSchool(), "waiting for exam")
+                .stream()
+                .map(UserRelationInfo::new)
+                .collect(Collectors.toList()));
+        return list;
     }
     
 }
